Extract asset matching into asset_matcher and split unit tests.
Move match_asset helpers out of IntentCompiler into a dedicated module
and add test_asset_matcher with shared fake-registry fixtures.
